requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
numpy
pydantic
lightrag-hku
raganything @ git+https://github.com/HKUDS/RAG-Anything.git
